refactor: Organize Bedrock provider into subdirectory structure
Move all Bedrock-specific provider files into provider/bedrock subdirectory
with a barrel export file for better code organization and modularity.

- Create provider/bedrock directory structure
- Move bedrock.ts → bedrock/provider.ts
- Move bedrockTypes.ts → bedrock/types.ts
- Move bedrockConverter.ts → bedrock/converter.ts
- Move test files to bedrock/ subdirectory
- Add barrel file (index.ts) for clean exports
- Update all import paths throughout codebase
- Apply Prettier formatting to documentation and tests

All 2728 tests passing (2703 core + 25 VSCode)
ESLint passing with 0 warnings
