+++ packages/core/src/core/openaiContentGenerator/provider/bedrock/types.ts
