+++ packages/core/src/core/openaiContentGenerator/provider/index.ts
@@ -3,7 +3,7 @@ export { DashScopeOpenAICompatibleProvider } from './dashscope.js';
 export { DeepSeekOpenAICompatibleProvider } from './deepseek.js';
 export { OpenRouterOpenAICompatibleProvider } from './openrouter.js';
 export { DefaultOpenAICompatibleProvider } from './default.js';
-export { BedrockOpenAICompatibleProvider } from './bedrock.js';
+export { BedrockOpenAICompatibleProvider } from './bedrock/index.js';
 export type {
   OpenAICompatibleProvider,
   DashScopeRequestMetadata,
